rit build
