--- Dockerfile
@@ -1,131 +1,0 @@
-ARG BUILD_IMAGE=adoptopenjdk/openjdk11-openj9:x86_64-alpine-jdk-11.0.5_10_openj9-0.17.0-slim
-FROM $BUILD_IMAGE
-
-# to run go things
-RUN apk add --no-cache \
-		ca-certificates
-
-ENV GOLANG_VERSION 1.14.1
-
-RUN set -eux; \
-	apk add --no-cache --virtual .build-deps \
-		bash \
-		gcc \
-		musl-dev \
-		openssl \
-		go \
-	; \
-	export \
-# set GOROOT_BOOTSTRAP such that we can actually build Go
-		GOROOT_BOOTSTRAP="$(go env GOROOT)" \
-# ... and set "cross-building" related vars to the installed system's values so that we create a build targeting the proper arch
-# (for example, if our build host is GOARCH=amd64, but our build env/image is GOARCH=386, our build needs GOARCH=386)
-		GOOS="$(go env GOOS)" \
-		GOARCH="$(go env GOARCH)" \
-		GOHOSTOS="$(go env GOHOSTOS)" \
-		GOHOSTARCH="$(go env GOHOSTARCH)" \
-	; \
-# also explicitly set GO386 and GOARM if appropriate
-# https://github.com/docker-library/golang/issues/184
-	apkArch="$(apk --print-arch)"; \
-	case "$apkArch" in \
-		armhf) export GOARM='6' ;; \
-		armv7) export GOARM='7' ;; \
-		x86) export GO386='387' ;; \
-	esac; \
-	\
-	wget -O go.tgz "https://golang.org/dl/go$GOLANG_VERSION.src.tar.gz"; \
-	echo '2ad2572115b0d1b4cb4c138e6b3a31cee6294cb48af75ee86bec3dca04507676 *go.tgz' | sha256sum -c -; \
-	tar -C /usr/local -xzf go.tgz; \
-	rm go.tgz; \
-	\
-	cd /usr/local/go/src; \
-	./make.bash; \
-	\
-	rm -rf \
-# https://github.com/golang/go/blob/0b30cf534a03618162d3015c8705dd2231e34703/src/cmd/dist/buildtool.go#L121-L125
-		/usr/local/go/pkg/bootstrap \
-# https://golang.org/cl/82095
-# https://github.com/golang/build/blob/e3fe1605c30f6a3fd136b561569933312ede8782/cmd/release/releaselet.go#L56
-		/usr/local/go/pkg/obj \
-	; \
-	apk del .build-deps; \
-	\
-	export PATH="/usr/local/go/bin:$PATH"; \
-	go version
-
-ENV GOPATH /go
-ENV PATH $GOPATH/bin:/usr/local/go/bin:$PATH
-
-RUN mkdir -p "$GOPATH/src" "$GOPATH/bin" && chmod -R 777 "$GOPATH"
-WORKDIR $GOPATH
-
-# to run aws things
-
-RUN apk update && apk add \
-    alpine-sdk \
-    gettext \
-    curl \
-    python \
-    py-pip \
-    jq \
-    git \
-    zip \
-    && pip install --no-cache-dir awscli==1.16.310 \
-    && apk del py-pip \
-    && rm -rf /var/cache/apk/* /root/.cache/pip/* /usr/lib/python2.7/site-packages/awscli/examples
-
-# to run node things
-
-ENV VERSION=v13.11.0 NPM_VERSION=6 YARN_VERSION=latest
-
-RUN apk upgrade --no-cache -U && \
-  apk add --no-cache curl make gcc g++ python linux-headers binutils-gold gnupg libstdc++ xz
-
-RUN for server in ipv4.pool.sks-keyservers.net keyserver.pgp.com ha.pool.sks-keyservers.net; do \
-    gpg --keyserver $server --recv-keys \
-      4ED778F539E3634C779C87C6D7062848A1AB005C \
-      B9E2F5981AA6E0CD28160D9FF13993A75599653C \
-      94AE36675C464D64BAFA68DD7434390BDBE9B9C5 \
-      B9AE9905FFD7803F25714661B63B535A4C206CA9 \
-      77984A986EBC2AA786BC0F66B01FBB92821C587A \
-      71DCFD284A79C3B38668286BC97EC7A07EDE3FC1 \
-      FD3A5288F042B6850C66B31F09FE44734EB7990E \
-      8FCCA13FEF1D0C2E91008E09770F7A9A5AE15600 \
-      C4F0DFFF4E8C1A8236409D08E73BC641CC11F4C8 \
-      DD8F2338BAE7501E3DD5AC78C273792F7D83545D \
-      A48C2BEE680E841632CD4E44F07496B3EB3C1762 && break; \
-  done
-
-RUN curl -sfSLO https://nodejs.org/dist/${VERSION}/node-${VERSION}.tar.xz && \
-  curl -sfSL https://nodejs.org/dist/${VERSION}/SHASUMS256.txt.asc | gpg -d -o SHASUMS256.txt && \
-  grep " node-${VERSION}.tar.xz\$" SHASUMS256.txt | sha256sum -c | grep ': OK$' && \
-  echo ${VERSION} &&\
-  ls -lah &&\
-  tar -xf node-${VERSION}.tar.xz && \
-  cd node-${VERSION} && \
-  ./configure --prefix=/usr ${CONFIG_FLAGS} && \
-  make -j$(getconf _NPROCESSORS_ONLN) && \
-  make install
-
-RUN if [ -z "$CONFIG_FLAGS" ]; then \
-    if [ -n "$NPM_VERSION" ]; then \
-      npm install -g npm@${NPM_VERSION}; \
-    fi; \
-    find /usr/lib/node_modules/npm -type d \( -name test -o -name .bin \) | xargs rm -rf; \
-    if [ -n "$YARN_VERSION" ]; then \
-      for server in ipv4.pool.sks-keyservers.net keyserver.pgp.com ha.pool.sks-keyservers.net; do \
-        gpg --keyserver $server --recv-keys \
-          6A010C5166006599AA17F08146C2130DFD2497F5 && break; \
-      done && \
-      curl -sfSL -O https://yarnpkg.com/${YARN_VERSION}.tar.gz -O https://yarnpkg.com/${YARN_VERSION}.tar.gz.asc && \
-      gpg --batch --verify ${YARN_VERSION}.tar.gz.asc ${YARN_VERSION}.tar.gz && \
-      mkdir /usr/local/share/yarn && \
-      tar -xf ${YARN_VERSION}.tar.gz -C /usr/local/share/yarn --strip 1 && \
-      ln -s /usr/local/share/yarn/bin/yarn /usr/local/bin/ && \
-      ln -s /usr/local/share/yarn/bin/yarnpkg /usr/local/bin/ && \
-      rm ${YARN_VERSION}.tar.gz*; \
-    fi; \
-  fi
-
-RUN npm install -g pkg
