+++ aws/create/bucket/config.json
@@ -1,6 +1,4 @@
 {
-  "command": "list",
-  "description": "List bucket aws",
   "inputs": [
     {
       "name" : "access_key",
@@ -19,6 +17,11 @@
         "qty" : 6,
         "newLabel" : "Type new value. "
       }
+    },
+    {
+      "name": "bucket",
+      "type": "text",
+      "label": "Type bucket name: "
     }
   ]
 }
