+++ aws/delete/bucket/help.txt
@@ -1,0 +1,1 @@
+Delete bucket aws
