+++ aws/delete/bucket/src/pkg/bucket/bucket.go
@@ -7,23 +7,13 @@ import (
 	"github.com/aws/aws-sdk-go/aws/credentials"
 	"github.com/aws/aws-sdk-go/aws/session"
 	"github.com/aws/aws-sdk-go/service/s3"
-	"github.com/jedib0t/go-pretty/table"
-	"os"
 )
 
 type Inputs struct {
 	Key        string
 	Secret     string
 	Region     string
-	BucketName string
-	Command    string
 }
-
-const (
-	list   = "list"
-	delete = "delete"
-	create = "create"
-)
 
 func (in Inputs) Run() {
 	if in.Key == "" || in.Secret == "" {
@@ -39,41 +29,7 @@ func (in Inputs) Run() {
 		return
 	}
 	svc := s3.New(sess)
-	switch in.Command {
-	case list:
-		in.runList(svc)
-	case create:
-		in.runCreate(svc)
-	case delete:
-		in.runDelete(svc)
-	default:
-		fmt.Printf("Command (%s) not found\n", in.Command)
-	}
-}
-
-func (in Inputs) runList(svc *s3.S3) {
-	res, err := in.list(svc)
-	if err != nil {
-		fmt.Printf("Error list bucket, error: %v", err)
-	} else {
-		printList(res)
-	}
-}
-
-func (in Inputs) runCreate(svc *s3.S3) {
-	if in.BucketName == "" {
-		fmt.Println("Not received bucket name!")
-		return
-	}
-	bn := &s3.CreateBucketInput{
-		Bucket: aws.String(in.BucketName),
-	}
-	_, err := svc.CreateBucket(bn)
-	if err != nil {
-		fmt.Printf("Failed create bucket. error: %v\n", err)
-		return
-	}
-	fmt.Println("Bucket created")
+	in.runDelete(svc)
 }
 
 func (in Inputs) runDelete(svc *s3.S3) {
@@ -114,17 +70,4 @@ func (in Inputs) list(svc *s3.S3) (*s3.ListBucketsOutput, error) {
 		return nil, fmt.Errorf("Failed list bucket, error: %s\n", err.Error())
 	}
 	return result, nil
-}
-
-func printList(r *s3.ListBucketsOutput) {
-	t := table.NewWriter()
-	t.SetOutputMirror(os.Stdout)
-	t.AppendHeader(table.Row{"Count", "Bucket Name", "Creation"})
-	for i, b := range r.Buckets {
-		t.AppendRows([]table.Row{
-			{i+1, aws.StringValue(b.Name), aws.TimeValue(b.CreationDate)},
-		})
-	}
-	t.SetStyle(table.StyleLight)
-	t.Render()
 }
