+++ aws/delete/bucket/src/pkg/prompt/prompt.go
