+++ aws/delete/help.txt
@@ -1,0 +1,1 @@
+Delete aws objects
