+++ aws/help.txt
@@ -1,0 +1,1 @@
+Apply Aws objects
