+++ aws/list/bucket/src/go.mod
@@ -1,0 +1,13 @@
+module aws/bucket
+
+go 1.14
+
+require (
+	github.com/aws/aws-sdk-go v1.32.2
+	github.com/fatih/color v1.9.0
+	github.com/go-openapi/strfmt v0.19.5 // indirect
+	github.com/gosuri/uitable v0.0.4
+	github.com/jedib0t/go-pretty v4.3.0+incompatible
+	github.com/manifoldco/promptui v0.7.0
+	github.com/mattn/go-runewidth v0.0.9 // indirect
+)
