+++ aws/list/help.txt
@@ -1,0 +1,1 @@
+List aws objects
