--- unzip-bin-configs.sh
@@ -1,2 +1,0 @@
-#!/bin/sh
-find formulas -name "*.zip" | while read filename; do unzip -o -d "`dirname "$filename"`" "$filename"; rm -f "$filename"; done;
